requirements.txt
fastapi
tensorflow
numpy
uvicorn
pillow
pydantic
scikit-learn